Approach_1/requirements.txt
google-generativeai
langchain
PyPDF2
faiss-cpu
streamlit
streamlit_chat